Added the README.md file, as well as changed the file hierarchy so that it is a bit more organized.
For the README.md file, I simply added my logo in ASCII and a single line of text saying that the project is still under constrution (no one should really be seeing it anyways since I have the project marked as private).

For the file hierarchy, I moved all of the web files (html, javascript, css, images, etc.) into a new folder 'src.' The reason for this is 2-fold. One, it makes it easier for me to test out a new idea for the website without having to create a new branch for it by simply copying and pasting the folder, and two, although Intellij has not complained about how the project is organized, this is how Intellij would create it by default, so I changed it to represent that.