FINALLY FIXED THE FIXED NAVBAR BUG!!! Also did some general code cleanup and edited the hover for buttons.
So apparently the navbar bug was caused by the transform: rotate(-0.0000000001deg) somehow. Deleting the line fixed the navbar. I don't know why but, honestly, I don't care as this bug has been plaguing me for the better part of the last few months. Now that it is fixed, I can finally get some real work done on this website.

 Also, I removed the slight color change on hover for the outlined buttons and replaced them with a transform: scale attribute.